LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_memory_cached.sub-expected.txt

Blocked access to external URL http://www1.localhost:8801/resource-timing/resources/blue.png?id=cached
Blocked access to external URL http://www1.localhost:8801/resource-timing/resources/blue.png?id=cached
Blocked access to external URL http://www1.localhost:8801/resource-timing/resources/blue.png?id=cached
Description

This test validates that a memory cached resource appears in the buffer once.


Harness Error (TIMEOUT), message = null

PASS There should be only one entry 
PASS http://localhost:8800/resource-timing/resources/blue.png?id=cached is expected to have initiatorType img 
PASS requestStart should be non-zero on the same-origin request 
PASS responseEnd should not be before startTime 
PASS duration should not be negative 
  
